handle auth errors and rethrow after logging and managing whatever is needed after error happens + add border indicator for refresh circle when the circle is dragged all the way down which indicates the refresh will happen if let go. + rename getAuthHeader to createAuthHeader in authutils + call revoke user endpoint when logging out in authservice

+++ service/AuthService.ts
@@ -3,8 +3,10 @@ import { decodeJwt } from "jose";
 
 import * as api from "@/api";
 import { appID } from "@/fotofamExtra.json";
-import { getAuthHeader, TokenClaims } from "@/util/auth";
+import { createAuthHeader, TokenClaims } from "@/util/auth";
 import { SECURE_STORE_VARS } from "@/util/constants";
+import { UserState } from "@/reducer/userReducer";
+import { Action, State, Store } from "@/reducer";
 
 export type AuthRequest = {
     username: string;
@@ -12,11 +14,13 @@ export type AuthRequest = {
 };
 
 export default class AuthService {
-    private dispatch: any;
+    private state: UserState;
+    private dispatch: React.Dispatch<Action>;
     private authToken: string | null = null;
     private baseUrl: string = "http://192.168.0.22:8082"; // use env var in real scenario. start using dotenv files
 
-    constructor([_, dispatch]: any, authToken: string | null) {
+    constructor([state, dispatch]: Store, authToken: string | null) {
+        this.state = state.user;
         this.dispatch = dispatch;
         this.authToken = authToken;
     }
@@ -26,22 +30,30 @@ export default class AuthService {
     }
 
     async authenticate(body: AuthRequest) {
-        const {
-            data: { token: at },
-        } = await api.post(`${this.baseUrl}/authenticate`, {
-            ...body,
-            applicationId: appID,
-        });
+        try {
+            const {
+                data: { token: at },
+            } = await api.post(`${this.baseUrl}/authenticate`, body);
+    
+            const {
+                data: { token: rt },
+            } = await api.post(
+                `${this.baseUrl}/authorize`,
+                undefined,
+                createAuthHeader(at)
+            );
+    
+            return rt;
+        } catch (error: any) {
+           const response = error.response;
+            const path = response.request.responseURL;
+            const status = response.status;
+            
+            console.error(path);
+            console.error(status);
 
-        const {
-            data: { token: rt },
-        } = await api.post(
-            `${this.baseUrl}/authorize`,
-            undefined,
-            getAuthHeader(at)
-        );
-
-        return rt;
+            throw error;
+        }
     }
 
     async login(rt: string) {
@@ -51,7 +63,7 @@ export default class AuthService {
             } = await api.post(
                 `${this.baseUrl}/login`,
                 undefined,
-                getAuthHeader(rt)
+                createAuthHeader(rt)
             );
 
             const decodedToken: TokenClaims = decodeJwt(token);
@@ -68,17 +80,36 @@ export default class AuthService {
 
             await setItemAsync(SECURE_STORE_VARS.authToken, token);
         } catch (error: any) {
-            // console.log(error.response.config.url);
-            // depending on which endpoint is matched here and what the status code is the app will decide on error message and / or clear app state and then show login screen again.
-            console.log(error.response.request.responseURL);
+            const response = error.response;
+            const path = response.request.responseURL;
+            const status = response.status;
+            
+            console.error(path);
+            console.error(status);
+            
+            if (status === 401 || status === 403) {
+                await deleteItemAsync(SECURE_STORE_VARS.authToken);
+            }
+
+            throw error;
         }
     }
 
     async logout() {
-        this.dispatch({
-            type: "LOGOUT_USER",
-        });
+        try {
+            this.dispatch({
+                type: "LOGOUT_USER",
+            });
 
-        await deleteItemAsync(SECURE_STORE_VARS.authToken);
+            await deleteItemAsync(SECURE_STORE_VARS.authToken);
+            await api.post(
+                `${this.baseUrl}/revoke/${this.state.username}`,
+                undefined
+            );
+        } catch (error: any) {
+            // console.log(error.response.config.url);
+            // depending on which endpoint is matched here and what the status code is the app will decide on error message and / or clear app state and then show login screen again.
+            console.log(error.response.request.responseURL);
+        }
     }
 }
